requirements.txt
requests
python_dotenv
pytest
pytest-mock
pytest-cov
typer[all]